[Feature] Add a button to reset all inputs to default
